Домашна 1/Проект/crypto-project/requirements.txt
aiohttp>=3.9.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0

